+++ python/interpret_text/experimental/common/preprocessor/glove_preprocessor.py
@@ -1,7 +1,7 @@
 import numpy as np
 import pandas as pd
 from typing import Any, Iterable, List
-from interpret_text.common.base_explainer import BaseTextPreprocessor
+from interpret_text.experimental.common.base_explainer import BaseTextPreprocessor
 
 PAD = "<PAD>"
 UNK = "<UNK>"
