+++ python/interpret_text/experimental/common/structured_model_mixin.py
