+++ python/interpret_text/experimental/introspective_rationale/introspective_rationale_components.py
@@ -7,7 +7,7 @@ import torch.nn as nn
 from torch.autograd import Variable
 from tqdm import tqdm
 
-from interpret_text.common.utils_introspective_rationale import generate_data
+from interpret_text.experimental.common.utils_introspective_rationale import generate_data
 
 
 class ClassifierWrapper():
